feat!: re-structure state and attributes (#7)
This changes the `state` to be the actual `timestamp` of the next non-canceled departure including delay. This allows to use the `state` programmatically, e.g. to compute a `timedelta` from it. Also, the sensor will be shown natively as a relative time, such as "in 5 minutes".

This also changes the `attributes` to allow for a list of `connections` to be returned.

BREAKING CHANGE: The `state` is now a timestamp and the attributes `products` and `time` have been renamed.